requirements.txt
django
djangorestframework
assertpy

